Update README.md
write a c++ program that will carry out the following tasks to demonstrate the concept if inheritance:
i)Define a base class named Shape with the following: A protected data member color(string) to represent the color of the shape.A public membee function getColor() to return the color of the shape.
ii)Define a derived class Rectangle that  inherits the Shape class and contains the following:Private data members length and width (both floats).A constructor that initializes length,width, and color by calling the base class constructor.A public member function area() that calculates and returns the area of the rectangle.
iii)Define a derived class *Circle* that inherits fron the *Shape* class and contains the following.A private data members radius (float). A constructor that intializes radius and function area() that calculates and returns the area of the circle using the formula π r^2.
iv)In the main()  function,perform the following actions:Create a Rectangle object with length-15, width-10, and color- ''blue''.Create a Circle object witb radius-2 and color-''pink''.Display the color and area of both the rectangle and the circle.